+++ server/quotes_service/get_factor.js
