+++ server/quotes_service/get_health_class.js
