+++ server/quotes_service/get_quote.js
@@ -1,9 +1,9 @@
-const find_health_class = require("./find_health_class")
-const find_factor =require("./find_factor")
+const get_health_class = require("./get_health_class")
+const get_factor =require("./get_factor")
 
 const getQuote = (term, coverage, age, height, weight)=>{
-    const healthClass = find_health_class.getHealthClass(height, weight)
-    const factor = find_factor.getFactor(healthClass, coverage, term, age)
+    const healthClass = get_health_class.getHealthClass(height, weight)
+    const factor = get_factor.getFactor(healthClass, coverage, term, age)
     const price = coverage / 1000 * factor
     return {
         price:price.toString(),
